requirements.txt
h5py==3.7.0
imutils==0.5.4
matplotlib==3.1.2
numpy==1.18.5
opencv_python_headless==4.5.5.64
pandas==1.4.2
Pillow==9.2.0
progress==1.6
pytest==4.6.9
scipy==1.3.3
setuptools==45.2.0
scikit-image==0.17.2
tensorboardX==2.5.1
torch==1.12.1
torchvision==0.13.1
